chore(core): move matching against handler to message handler abstraction
